NsId improvements
- Use a different hack when expanding macros so we don't allocate a new
  `NsId` for every identifier (!)

- Reduce the size of NsId to 32 bits. It's embedded in a lot of data
  structures and we shouldn't need more than 4 billion namespaces per
  compile unit

+++ src/hir/ns.rs
@@ -1,11 +1,11 @@
-use syntax::span::Span;
 use syntax::datum::Datum;
+use syntax::span::Span;
 
 #[derive(Clone, Copy, PartialEq, Eq, Hash, Debug)]
-pub struct NsId(usize);
+pub struct NsId(u32);
 
 impl NsId {
-    pub fn new(id: usize) -> NsId {
+    pub fn new(id: u32) -> NsId {
         NsId(id)
     }
 }
@@ -128,7 +128,7 @@ impl NsDatum {
 }
 
 pub struct NsIdAlloc {
-    curr_ns_id: usize,
+    curr_ns_id: u32,
 }
 
 impl NsIdAlloc {
